requirements.txt
django
djangorestframework
django-cors-headers
gemini
twilio
channels_redis
django-channels
requests
websockets